requirements.txt
aiogram>=3.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
deep-translator>=1.11.4
google-generativeai>=0.4.0